+++ src/quantization/scalar_quantization/bf16_quantizer_test.cpp
@@ -54,7 +54,7 @@ TestComputeMetricBF16(uint64_t dim, int count, float error = 1e-5) {
     auto allocator = SafeAllocator::FactoryDefaultAllocator();
     BF16Quantizer<metric> quantizer(dim, allocator.get());
     TestComputeCodes<BF16Quantizer<metric>, metric>(quantizer, dim, count, error);
-    TestComputer<BF16Quantizer<metric>, metric>(quantizer, dim, count, error, true, 1.0, 1.0);
+    TestComputer<BF16Quantizer<metric>, metric>(quantizer, dim, count, error, 1.0, true, 1.0, 1.0);
     // TODO(LHT): fix quantize error
 }
